+++ fantasy_cpc_display_controller/src/text_layer/text_layer.rs
@@ -1,21 +1,21 @@
 use crate::{config::*, color_palettes::*};
 
-const DEFAULT_COLOR: usize = WHITE;
-const DEFAULT_BKG_COLOR: usize = BLACK;
+const DEFAULT_COLOR: u8 = WHITE;
+const DEFAULT_BKG_COLOR: u8 = BLACK;
 
 #[derive(Clone, Copy)]
 pub struct TextLayerChar {
     pub c: char,
-    pub color: usize,
-    pub bkg_color: usize,
+    pub color: u8,
+    pub bkg_color: u8,
     pub swap: bool,
     pub blink: bool,
     pub shadowed: bool
 }
 
 pub struct TextLayer {
-    pub default_color: usize,
-    pub default_bkg_color: usize,
+    pub default_color: u8,
+    pub default_bkg_color: u8,
     char_map: [Option<TextLayerChar>; TEXT_COLUMNS * TEXT_ROWS],
 }
 
@@ -65,12 +65,12 @@ impl TextLayer {
     }
 
     /// Inserts a character in the char_map at the specified index.
-    pub fn insert_char(&mut self, index: usize, c: char, color: Option<usize>, bkg_color: Option<usize>, swap: bool, blink: bool, shadowed: bool) {
+    pub fn insert_char(&mut self, index: usize, c: char, color: Option<u8>, bkg_color: Option<u8>, swap: bool, blink: bool, shadowed: bool) {
         self.insert_text_layer_char(index, TextLayerChar {c, color: color.unwrap_or(DEFAULT_COLOR), bkg_color: bkg_color.unwrap_or(DEFAULT_BKG_COLOR), swap, blink, shadowed});
     }
 
     /// Inserts a character in the char_map at the specified x and y position.
-    pub fn insert_char_xy(&mut self, x: usize, y: usize, c: char, color: Option<usize>, bkg_color: Option<usize>, swap: bool, blink: bool, shadowed: bool) {
+    pub fn insert_char_xy(&mut self, x: usize, y: usize, c: char, color: Option<u8>, bkg_color: Option<u8>, swap: bool, blink: bool, shadowed: bool) {
         let index = text_coord_to_index(x, y);
         self.insert_char(index, c, color, bkg_color, swap, blink, shadowed);
     }
@@ -82,7 +82,7 @@ impl TextLayer {
     }
 
     /// Inserts a string in the char_map at the specified index.
-    pub fn insert_string(&mut self, index: usize, string: &str, color: Option<usize>, bkg_color: Option<usize>, swap: bool, blink: bool, shadowed: bool) {
+    pub fn insert_string(&mut self, index: usize, string: &str, color: Option<u8>, bkg_color: Option<u8>, swap: bool, blink: bool, shadowed: bool) {
         if !string.is_empty() {
             for (char_count, c) in string.chars().enumerate() {
                 self.insert_char(index + char_count, c, color, bkg_color, swap, blink, shadowed);
@@ -91,7 +91,7 @@ impl TextLayer {
     }
 
     /// Inserts a string in the char_map at the specified x and y position.
-    pub fn insert_string_xy(&mut self, x: usize, y: usize, string: &str, color: Option<usize>, bkg_color: Option<usize>, swap: bool, blink: bool, shadowed: bool) {
+    pub fn insert_string_xy(&mut self, x: usize, y: usize, string: &str, color: Option<u8>, bkg_color: Option<u8>, swap: bool, blink: bool, shadowed: bool) {
         let index = text_coord_to_index(x, y);
         self.insert_string(index, string, color, bkg_color, swap, blink, shadowed);
     }
