+++ src/apps/weather_app.rs
@@ -317,12 +317,12 @@ impl WeatherApp {
 
     fn generate_cloud() -> Cloud {
         //Max box size containing circle centers
-        let box_width: isize = 64;
-        let box_height: isize = 16;
+        let box_width: i32 = 64;
+        let box_height: i32 = 16;
 
         //Min and Max cloud circles
-        let min_circle_r: usize = 8;
-        let max_circle_r: usize = 32;
+        let min_circle_r: u32 = 8;
+        let max_circle_r: u32 = 32;
 
         //Generate 5 random circles with random radius within box
         let mut random = rand::rng();
@@ -330,14 +330,14 @@ impl WeatherApp {
 
         for circle in circles.iter_mut().enumerate() {
             circle.1 .0 = random.random_range(
-                ((box_width / 5) * circle.0 as isize)..((box_width / 5) * (circle.0 as isize + 1)),
-            );
-            circle.1 .1 = random.random_range(0..box_height);
-            circle.1 .2 = random.random_range(min_circle_r..=max_circle_r);
+                ((box_width / 5) * circle.0 as i32)..((box_width / 5) * (circle.0 as i32 + 1)),
+            ) as isize;
+            circle.1 .1 = random.random_range(0..box_height) as isize;
+            circle.1 .2 = random.random_range(min_circle_r..=max_circle_r) as usize;
         }
 
-        let cloud_x = random.random_range(-300..-100);
-        let cloud_y = random.random_range(0..80);
+        let cloud_x = random.random_range(-300 as i32..-100 as i32) as isize;
+        let cloud_y = random.random_range(0 as u32..80 as u32) as isize;
 
         Cloud {
             circles,
